infrastructure/web-apps/blog-app/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
redis==4.6.0